More separation of concerns in main script.
