+++ src/main.js
@@ -4,6 +4,10 @@ import { ClickAndKeyControls } from './config/clickControlHelper.js';
 
 import { drawImage } from './drawing/drawImage.js';
 
+import { loadDataSource, pixelToWorldUnits, prepareDrawingContext, drawHelpers } from './config/utils.js';
+
+import { loadThenDraw } from './config/loadThenDraw.js';
+
 import { OutlineEffect } from 'outline-effect';
 
 import { usePanoramicCubeBackground, useProceduralBackground, usePanoramicCubeBackgroundSixFaces } from './drawing/drawBackground.js';
@@ -27,24 +31,6 @@ const DEBUG = false;
 
 
 const textureLoader = new TextureLoader();
-const fileLoader = new FileLoader();
-
-
-const loadDataSource = (scene, dataSrc, drawFunc, state) => {
-    const jsonPath = `./data/${dataSrc}.json`;
-
-    fileLoader.load(
-        jsonPath,
-        (dataString) => {
-            const data = JSON.parse(dataString);
-            drawFunc(scene, data, state);
-        },
-        undefined, // onProgress
-        (err) => {
-            console.error(`Error loading ${jsonPath}`, err);
-        }
-    );
-}
 
 const THREEJS_DRAWINGS = {
     'room': () => import('./drawing/drawRoom.js').then(m => m.roomDrawing),
@@ -86,33 +72,6 @@ const THREEJS_DRAWINGS = {
 };
 
 
-function drawHelpers(scene, threejsDrawing) {
-    const refGeometry = new BoxGeometry(1, 1, 1); // 1x1x1 cube
-    const refMaterial = new MeshNormalMaterial({ wireframe: true });
-    const refCube = new Mesh(refGeometry, refMaterial);
-    refCube.position.set(0, 0.5, 0); // sit on ground
-    scene.add(refCube);
-
-    const gridHelper = new GridHelper(10, 10); // 10x10 units
-    scene.add(gridHelper);
-
-    //window.debugObject = object; // now accessible from console
-
-//    document.addEventListener('keydown', (e) => {
-//        if (e.key === 'ArrowUp') window.debugObject.position.z -= 0.1;
-//        if (e.key === 'ArrowDown') window.debugObject.position.z += 0.1;
-//        if (e.key === 'ArrowLeft') window.debugObject.position.x -= 0.1;
-//        if (e.key === 'ArrowRight') window.debugObject.position.x += 0.1;
-//    });
-}
-
-function pixelToWorldUnits(pixelSize, distance, camera) {
-    const fovInRad = camera.fov * (Math.PI / 180);
-    const screenHeight = 2 * Math.tan(fovInRad / 2) * distance;
-    const pixelHeightInWorld = screenHeight / window.innerHeight;
-    return pixelSize * pixelHeightInWorld;
-}
-
 document.addEventListener('DOMContentLoaded', () => {
     const drawingName = document.querySelector('meta[name="threejs_drawing_name"]').content;
 
@@ -120,17 +79,6 @@ document.addEventListener('DOMContentLoaded', () => {
         contentLoadedCallback(threejsDrawing);
     })
 })
-
-async function prepareDrawingContext(threejsDrawing, scene, camera, renderer, controls, cssRenderer) {
-    Object.assign(threejsDrawing.data, {
-        scene,
-        camera,
-        renderer,
-        controls,
-        cssRenderer,
-    });
-    return threejsDrawing;
-}
 
 
 const CUBE_DEFS = [
@@ -160,43 +108,7 @@ async function contentLoadedCallback(threejsDrawing) {
 
     for (const {func, dataSrc, dataType} of threejsDrawing.drawFuncs) {
         if (dataSrc) {
-            const data_src = dataSelected ? dataSelected : dataSrc;
-            console.log(`Loading data source: ${data_src}`);
-            threejsDrawing.data.dataSrc = data_src;
-            if (dataType === 'svg') {
-                import('svgloader').then(m => {
-                    const svgLoader = new m.SVGLoader();
-                    svgLoader.load(`./imagery/${data_src}_out_annotated.svg`, (data) => {
-                        func(scene, data, threejsDrawing);
-                    });
-                });
-            } else if (dataType === 'json') {
-                const worldWidth = pixelToWorldUnits(480, 5, camera); // 480px at 5 units away
-                const worldHeight = pixelToWorldUnits(360, 5, camera);
-                threejsDrawing.data.worldWidth = worldWidth;
-                threejsDrawing.data.worldHeight = worldHeight;
-                loadDataSource(scene, data_src, func, threejsDrawing);
-            } else if (dataType === 'gltf') {
-                import('gltfloader').then(m => {
-                    const GLTFLoader = m.GLTFLoader;
-                    const gltfLoader = new GLTFLoader();
-                    gltfLoader.load(`./imagery/${data_src}.glb`, (gltf) => {
-                        console.log(`Loaded GLTF model: ${data_src}`, gltf);
-                        func(scene, gltf, threejsDrawing);
-                    });
-                });
-            } else if (dataType === 'exr') {
-                import('exrloader').then(m => {
-                    const EXRLoader = m.EXRLoader;
-                    const exrLoader = new EXRLoader();
-                    exrLoader.setDataType(FloatType).load(`./textures/${data_src}.exr`, (texture) => {
-                        console.log(`Loaded EXR texture: ${data_src}`, texture);
-                        func(scene, texture, threejsDrawing);
-                    });
-                });
-            } else {
-                console.error(`Unknown data type: ${dataType}`);
-            }
+            loadThenDraw(scene, func, dataSrc, dataType, camera, threejsDrawing, dataSelected);
         } else {
             await func(scene, threejsDrawing);
         }
